chore: switch jobs logic, move files
